photo REST controller #1

+++ src/main/java/com/archive/fursuit/controllers/PhotoREST.java
@@ -1,5 +1,6 @@
 package com.archive.fursuit.controllers;
 
+import com.archive.fursuit.Event;
 import com.archive.fursuit.Photo;
 import com.archive.fursuit.services.impl.EventServiceImpl;
 import com.archive.fursuit.services.impl.PhotoServiceImpl;
@@ -9,7 +10,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/photos")
+@RequestMapping("/api/")
 public class PhotoREST {
 
     @Autowired
@@ -18,9 +19,14 @@ public class PhotoREST {
     @Autowired
     private EventServiceImpl eventServiceImpl;
 
-    @GetMapping("find")
-    public List<Photo> getPhotos(){
-        return photoServiceImpl.findAllPhotos();
+    @GetMapping("public/photos/findAll")
+    public List<Photo> getPhotos() {
+        List<Photo> allPhotos = photoServiceImpl.findAllPhotos();
+        if (allPhotos.isEmpty()) {
+            return null;
+        } else {
+            return allPhotos;
+        }
     }
 
     @PostMapping("upload")
@@ -29,17 +35,46 @@ public class PhotoREST {
         return photo;
     }
 
-    @GetMapping("delete/{id}")
-    public void deletePhoto(@PathVariable ("id") long id){
-        photoServiceImpl.deletePhoto(id);
+    @DeleteMapping("private/photo/delete/{id}")
+    public String deletePhoto(@PathVariable ("id") long id){
+        Photo deletedPhoto = photoServiceImpl.getPhotoById(id);
+        if(deletedPhoto == null) {
+            return String.format("A photo with ID %s doesn't exist.", id);
+        }
+        else {
+            photoServiceImpl.deletePhoto(id);
+            return String.format("Photo with ID %s successfully deleted.", id);
+        }
     }
 
-//    @GetMapping("api/photo/assign/{eventId}/{photoId}")
-//    public Event assignPhoto(@PathVariable ("eventId") long eventId, @PathVariable ("photoId") long photoId){
-//        Event event = eventServiceImpl.getEventById(eventId);
-//        Photo photo = photoServiceImpl.getPhotoById(photoId);
-//        photoServiceImpl.assignEvent(photo, eventId);
-//        return event;
-//    }
+    @PostMapping("private/photo/move/{photoId}/{eventId}")
+    public String movePhoto(@PathVariable ("photoId") long photoId, @PathVariable ("eventId") long eventId){
+        Photo movedPhoto = photoServiceImpl.getPhotoById(photoId);
+        if(movedPhoto == null) {
+            return String.format("A photo with ID %s doesn't exist.", photoId);
+        }
+        else {
+            return photoServiceImpl.assignEvent(movedPhoto, eventId);
+        }
+    }
 
+    @PutMapping("private/photo/edit/{id}")
+    public Photo editPhoto(@PathVariable ("id") long id, @RequestBody Photo photo){
+        Photo editedPhoto = photoServiceImpl.getPhotoById(id);
+        if(editedPhoto == null){
+            return null;
+        }
+        else {
+            editedPhoto.setLabel(photo.getLabel());
+            editedPhoto.setPhotographer(photo.getPhotographer());
+            editedPhoto.setDate(photo.getDate());
+            photoServiceImpl.savePhoto(editedPhoto);
+            return editedPhoto;
+        }
+    }
+
+
+    //downloadPhoto
+
+    //uploadPhoto with image
 }
